App using API
The App uses the API now
app.pomodoro.nu is now used to get to the app
www.pomodoro.nu gets a project webpage (wwwIndex.scala)
a workaround for cross domain scripting is in use in routes
<main> is used through the API
html5shiv used for olderbrowsers with no support for <main>
app css is now app.css. main.css is for webpage.
upgraded to jQuery 1.9.0